feat: add closed view when ticket closed by noch fragen
this is a bit problematic because when auto closing, we don't have access to the interaction so this required some other changes as well. that is why the user clicking "no questions left" uses the same `close_ticket` function (which needed a few changes to make it compatible), whereas the automatic one needs a completely different flow.